Move the memory allocator to the IO module
