+++ community/io/src/main/java/org/neo4j/io/mem/MemoryAllocator.java
@@ -17,7 +17,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.neo4j.unsafe.impl.internal.dragons;
+package org.neo4j.io.mem;
 
 /**
  * A MemoryAllocator is simple: it only allocates memory, until it itself is finalizable and frees it all in one go.
