+++ community/io/src/test/java/org/neo4j/io/mem/MemoryAllocatorTest.java
@@ -17,9 +17,11 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.neo4j.unsafe.impl.internal.dragons;
+package org.neo4j.io.mem;
 
 import org.junit.Test;
+
+import org.neo4j.unsafe.impl.internal.dragons.UnsafeUtil;
 
 import static org.hamcrest.Matchers.is;
 import static org.hamcrest.Matchers.not;
